feat!: Response.filter_header(), with_headers(), HeaderError
In https://github.com/tiny-http/tiny-http/issues/209 it has been requested to provide a feature to remove the Server HTTP header from responses.

You can add now a filter to the response with filter_header(HeaderField).
This prevents the header to be sent with the response.

All methods adding header stuff to the response now return a Result with HeaderError if the header got not added.

For adding multiple headers at once, there is now add_headers(Vec) and with_headers(Vec).