refactor!: Header, HeaderField, HeaderFieldValue, ConnectionHeader 1/3
All changes has been done with performance and safety in mindset.

The values in headers needs some stricter check than it has been with just ASCII before.
For this value isn't AsciiString anylonger but an own struct HeaderFieldValue.

The internal algorithms now are based on u8/byte arrays and slices, at least where it makes sense.
If there is an need to store changed data on the heap it makes more sense to stay on AsciiString types.
Although it needs to be checked, if there is an advantage of AsciiString over std String.
The Ascii limit is not enough to be safe and so nevertheless the bytes need principally a check.

The refactoring gets completed with 2 more commits...