feat!: correct HTTP/0.9 support
HTTP/0.9 doesn't know headers, content-types and anything different to a GET request.
If there is a HTTP/0.9 request, the provided request should be correct now, but you need to handle the response content yourself.
There are no status codes. The client parses the response like an HTML document and there has been a line limit of 80 chars.